refactor modules/typesense
