+++ src/modules/typesense/controllers/typesense.controller.ts
@@ -1,25 +1,23 @@
 import { Controller, Get, Query } from '@nestjs/common'
 import { type CollectionSchema } from 'typesense/lib/Typesense/Collection.js'
-import { ApiOAuth2, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger'
+import { ApiBearerAuth, ApiResponse, ApiTags } from '@nestjs/swagger'
 import { MigrateTypesenseQueryDto } from '../dtos/migrate-typesense-query.dto.js'
 import { ImportTypesenseQueryDto } from '../dtos/import-typesense-query.dto.js'
-import { Permissions } from '../../permissions/permissions.decorator.js'
-import { Permission } from '../../permissions/permission.enum.js'
-import { TypesenseCollectionName } from '../enums/typesense-collection-index.enum.js'
+import { Permissions } from '../../permissions/decorators/permissions.decorator.js'
+import { Permission } from '../../permissions/enums/permission.enum.js'
 import { TypesenseInitializationService } from '../services/typesense-initialization.service.js'
+import { getCollectionsTypenseResponse, importTypesenseResponse, migrateTypesenseResponse } from '../docs/typesense-response.docs.js'
 
 @ApiTags('Typesense')
 @Controller('typesense')
-@ApiOAuth2([])
+@ApiBearerAuth()
 export class TypesenseController {
-  constructor (private readonly typesenseImportService: TypesenseInitializationService) {}
-  @Get('migrate')
-  @ApiQuery({ required: false, name: 'fresh', type: 'boolean' })
-  @ApiQuery({ required: false, name: 'collections', enum: TypesenseCollectionName, isArray: true })
-  @ApiResponse({
-    status: 200,
-    description: 'Successfully migrated collections'
-  })
+  constructor (
+    private readonly typesenseImportService: TypesenseInitializationService
+  ) {}
+
+  @Get('/migrate')
+  @ApiResponse(migrateTypesenseResponse)
   @Permissions(Permission.ADMIN)
   async migrate (
     @Query() query: MigrateTypesenseQueryDto
@@ -27,12 +25,8 @@ export class TypesenseController {
     await this.typesenseImportService.migrate(query.fresh, query.collections)
   }
 
-  @Get('import')
-  @ApiQuery({ required: false, name: 'collections', enum: TypesenseCollectionName, isArray: true })
-  @ApiResponse({
-    status: 200,
-    description: 'Successfully imported collections'
-  })
+  @Get('/import')
+  @ApiResponse(importTypesenseResponse)
   @Permissions(Permission.ADMIN)
   async import (
     @Query() query: ImportTypesenseQueryDto
@@ -40,11 +34,8 @@ export class TypesenseController {
     await this.typesenseImportService.import(query.collections)
   }
 
-  @Get('collections')
-  @ApiResponse({
-    status: 200,
-    description: 'Successfully returned collections'
-  })
+  @Get('/collections')
+  @ApiResponse(getCollectionsTypenseResponse)
   @Permissions(Permission.ADMIN)
   async getCollections (): Promise<CollectionSchema[]> {
     return await this.typesenseImportService.retrieveCollections()
