+++ src/modules/typesense/docs/typesense-response.docs.ts
@@ -1,0 +1,16 @@
+import { type ApiResponseOptions } from '@nestjs/swagger'
+
+export const migrateTypesenseResponse: ApiResponseOptions = {
+  status: 200,
+  description: 'Successfully migrated collections.'
+}
+
+export const importTypesenseResponse: ApiResponseOptions = {
+  status: 200,
+  description: 'Successfully imported collections.'
+}
+
+export const getCollectionsTypenseResponse: ApiResponseOptions = {
+  status: 200,
+  description: 'Successfully retrieved collections.'
+}
