+++ src/modules/typesense/dtos/migrate-typesense-query.dto.ts
@@ -1,20 +1,21 @@
 import {
   IsBoolean,
   IsEnum,
-  IsIn,
   IsOptional
 } from 'class-validator'
-import { Transform } from 'class-transformer'
+import { Type } from 'class-transformer'
+import { ApiProperty } from '@nestjs/swagger'
 import { TypesenseCollectionName } from '../enums/typesense-collection-index.enum.js'
 
 export class MigrateTypesenseQueryDto {
   @IsOptional()
   @IsBoolean()
-  @Transform(({ obj, key }) => obj[key] === 'true')
-  @IsIn([true, false])
-  fresh: boolean
+  @Type(() => Boolean)
+  @ApiProperty({ type: Boolean, description: 'Whether to refresh the Typesense collections.' })
+  fresh: boolean = false
 
   @IsOptional()
   @IsEnum(TypesenseCollectionName, { each: true })
+  @ApiProperty({ type: String, enum: TypesenseCollectionName, isArray: true, description: 'The collections to migrate.' })
   collections: TypesenseCollectionName[] = Object.values(TypesenseCollectionName)
 }
